Facade pattern (#2)
* Facade layer, interfaces added

* Facade implemented, utils added

---------
