+++ src/main/java/com/demoproject/cricketapp/utils/MatchUtils.java
@@ -1,0 +1,12 @@
+package com.demoproject.cricketapp.utils;
+
+import lombok.Data;
+import lombok.RequiredArgsConstructor;
+import lombok.experimental.UtilityClass;
+import org.springframework.stereotype.Component;
+
+
+@UtilityClass
+public class MatchUtils {
+
+}
